# move to PCMPESTRI, remove need for allocation and fix buffer overrun