--- a/parse_asm.s
+++ b/parse_asm.s
@@ -15,11 +15,13 @@ TEXT ·hasAsm(SB),$0-1
 RET
 
 TEXT ·scanNumberChars(SB),4,$0-40
-    // load range (0-9)
+    // load range 0-9
     MOVQ $0x000000FF3a2F01, BX
     MOVQ BX, X0
+    MOVQ $4, AX
+
     // first argument is the byte array we're scanning
-    MOVQ s_base+0(FP), SI
+    MOVQ s+0(FP), SI
     MOVQ offset+24(FP), BX
     ADDQ BX, SI
 
@@ -27,36 +29,33 @@ TEXT ·scanNumberChars(SB),4,$0-40
     // now load the string length into AX, loop control
     // we'll do one loop for every substring of up to 16 bytes,
     // thats (x + 15) >> 4
-	MOVQ s_len+8(FP), AX
-    SUBQ BX, AX
-    // save off number of bytes
-    MOVQ AX, DX
-    ADDQ $15, AX
-    SHRQ $4, AX
+    MOVQ s_len+8(FP), DX
+    SUBQ BX, DX
 
-	// SI holds the memory address of slice + offset
-    // AX holds the number of iterations required to process the entire array
+    // AX holds $6 - len of our range (argument to PCMPESTRI)
     // X0 holds numeric byte ranges that should cause us to halt scanning
+    // DX holds the number of bytes remaining (argument to PCMPESTRI)
+    // SI holds the memory address of slice + offset
+    // BX holds the number of bytes processed
 
-    TESTQ AX,AX
-    JZ scanNumberCharsEnd
     MOVQ $0, BX
+    TESTQ DX,DX
+    JZ scanNumberCharsEnd
 scanNumberCharsLoop:
-    // PCMPISTRI $0x24, 0(SI), X0
-    BYTE $0x66 ;BYTE $0x0F;	BYTE $0x3A;	BYTE $0x63 ; BYTE $0x06 ; BYTE $0x24
+    PCMPESTRI $0x04, 0(SI), X0
     JC scanNumberCharsFound
     ADDQ $16, SI
-    ADDQ $1, BX
-    DECQ AX
-    TESTQ AX,AX
-    JNZ scanNumberCharsLoop
+    ADDQ $16, BX
+    CMPQ DX, $16
+    SUBQ $16, DX
+    JA scanNumberCharsLoop
 
 scanNumberCharsEnd:
-    MOVQ DX, ret+32(FP)
+    ADDQ DX, BX
+    MOVQ BX, ret+32(FP)
     RET
 
 scanNumberCharsFound:
-    SHLQ $4, BX
     ADDQ CX, BX
     MOVQ BX, ret+32(FP)
     RET
@@ -65,6 +64,8 @@ TEXT ·scanNonSpecialStringChars(SB),4,$0-40
     // load range (control, '"', and '\')
     MOVQ $0x5c5c22221f01, BX
     MOVQ BX, X0
+    MOVQ $6, AX
+
     // first argument is the byte array we're scanning
     MOVQ s+0(FP), SI
     MOVQ offset+24(FP), BX
@@ -74,85 +75,33 @@ TEXT ·scanNonSpecialStringChars(SB),4,$0-40
     // now load the string length into AX, loop control
     // we'll do one loop for every substring of up to 16 bytes,
     // thats (x + 15) >> 4
-	MOVQ s_len+8(FP), AX
-    SUBQ BX, AX
-    // save off number of bytes
-    MOVQ AX, DX
-    ADDQ $15, AX
-    SHRQ $4, AX
+    MOVQ s_len+8(FP), DX
+    SUBQ BX, DX
 
-	// SI holds the memory address of slice + offset
-    // AX holds the number of iterations required to process the entire array
+    // AX holds $6 - len of our range (argument to PCMPESTRI)
     // X0 holds numeric byte ranges that should cause us to halt scanning
+    // DX holds the number of bytes remaining (argument to PCMPESTRI)
+    // SI holds the memory address of slice + offset
+    // BX holds the number of bytes processed
 
-    TESTQ AX,AX
-    JZ scanNonSpecialStringCharsEnd
     MOVQ $0, BX
+    TESTQ DX,DX
+    JZ scanNonSpecialStringCharsEnd
 scanNonSpecialStringCharsLoop:
-    // PCMPISTRI $0x24, 0(SI), X0
-    BYTE $0x66 ;BYTE $0x0F;	BYTE $0x3A;	BYTE $0x63 ; BYTE $0x06 ; BYTE $0x24
+    PCMPESTRI $0x04, 0(SI), X0
     JC scanNonSpecialStringCharsFound
     ADDQ $16, SI
-    ADDQ $1, BX
-    DECQ AX
-    TESTQ AX,AX
-    JNZ scanNonSpecialStringCharsLoop
+    ADDQ $16, BX
+    CMPQ DX, $16
+    SUBQ $16, DX
+    JA scanNonSpecialStringCharsLoop
 
 scanNonSpecialStringCharsEnd:
-    MOVQ DX, ret+32(FP)
-    RET
-
-scanNonSpecialStringCharsFound:
-    SHLQ $4, BX
-    ADDQ CX, BX
+    ADDQ DX, BX
     MOVQ BX, ret+32(FP)
     RET
 
-
-// NOTE: as implemented, the overhead of this routine for terse json outwieghs the
-// benefit
-TEXT ·scanWhitespaceChars(SB),4,$0-40
-    // load range (control, '"', and '\')
-    MOVQ $0xff211f0e0801, BX
-    MOVQ BX, X0
-    // first argument is the byte array we're scanning
-    MOVQ s+0(FP), SI
-    MOVQ offset+24(FP), BX
-    ADDQ BX, SI
-
-    // second argument is the offset into the array
-    // now load the string length into AX, loop control
-    // we'll do one loop for every substring of up to 16 bytes,
-    // thats (x + 15) >> 4
-	MOVQ s_len+8(FP), AX
-    SUBQ BX, AX
-    MOVQ AX, DX
-    ADDQ $15, AX
-    SHRQ $4, AX
-
-	// SI holds the memory address of slice + offset
-    // AX holds the number of iterations required to process the entire array
-    // X0 holds numeric byte ranges that should cause us to halt scanning
-
-    TESTQ AX,AX
-    JZ scanWhitespaceCharsEnd
-    MOVQ $0, BX
-scanWhitespaceCharsLoop:
-    // PCMPISTRI $0x24, 0(SI), X0
-    BYTE $0x66 ;BYTE $0x0F;	BYTE $0x3A;	BYTE $0x63 ; BYTE $0x06 ; BYTE $0x24
-    JC scanWhitespaceCharsFound
-    ADDQ $16, SI
-    ADDQ $1, BX
-    DECQ AX
-    TESTQ AX,AX
-    JNZ scanWhitespaceCharsLoop
-
-scanWhitespaceCharsEnd:
-    MOVQ DX, ret+32(FP)
-    RET
-
-scanWhitespaceCharsFound:
-    SHLQ $4, BX
+scanNonSpecialStringCharsFound:
     ADDQ CX, BX
     MOVQ BX, ret+32(FP)
     RET