PCMPISTRI requires implict string length (null padding).  buffer over-run was possible, and a half-brained mitigation was in place that was inefficient.  This moves to explicit length SSE scanning (PCMPESTRI) and removes the allocation.  The result is faster and more correct:

before:
```
$ go test -bench=".*" -run="XXX"
goos: linux
goarch: amd64
pkg: github.com/lloyd/goj/test
BenchmarkGojScanning-24        	     200	   7544816 ns/op	 257.19 MB/s
BenchmarkStdJSONScanning-24    	     100	  14643059 ns/op	 132.52 MB/s
```
after
```
$ go test -bench=".*" -run="XXX"
goos: linux
goarch: amd64
pkg: github.com/lloyd/goj/test
BenchmarkGojScanning-24        	     300	   4842530 ns/op	 400.71 MB/s
BenchmarkStdJSONScanning-24    	     100	  14554536 ns/op	 133.32 MB/s
```